Don't using logging when importing wok.renders.
Because logging has not been set up yet (because we haven't parsed CLI
options yet) making calls to logging.* will cause logging to be set up
with default values, which is bad.

This is a bad fix. The real fix should be to move parsing of CLI options
to another stage so that we can set up logging before importing the
renders. But that is a bigger change.

ugh ugh ugh ugh ugh ugh ugh.